+++ tests/test_shortcut.py
@@ -17,8 +17,23 @@ def test_loop_shortcut(model, expt_plain):
     assert loop_model(a=0, b=2, d=[1, 2], f=3) == ([8, 0], 1.0, 9)
 
     loop_expt = loop_shortcut(expt_plain, "f")
-    assert loop_expt(a=0, b=2, d=1, f=[3, 4]) ==  ((8, 9), 1.0, (16, 16))
+    print(loop_expt.graph.nodes['f_loop_node']['func'](c=2, e=1, f=[3, 4]))
+    assert loop_expt(a=0, b=2, d=1, f=[3, 4]) == ((8, 9), 1.0, (16, 16))
 
+
+def test_loop_shortcut_single_node(model, expt_plain):
+    """Test loop shortcut when the subgraph is a single node
+
+    The resulting value should create a looped model of d. For expt,
+    the d is already looped.
+    """
+
+    loop_model = loop_shortcut(loop_shortcut(model, "d"), "d")
+    assert loop_model(a=0, b=2, d=[[1, 2], [1, 2]], f=3) == ([[8, 0], [8, 0]], 1.0, 9)
+
+    loop_model = loop_shortcut(loop_shortcut(model, "f"), "f")
+    # the output is actually k1, p1, k2, p2
+    assert loop_model(a=0, b=2, d=1, f=[[3, 4], [3, 4]]) == ([(8, 9), (16, 16)], 1.0, [(8, 9), (16, 16)])
 
 def test_loop_shortcut_top_level(model, expt_plain):
     """Test loop shortcut when the parameter is used at the top level